+++ example/lib/main/main_page.dart
@@ -29,9 +29,6 @@ class _MainPageState extends State<MainPage> {
               ).contains(RouteConstants().badgeNavigation)
               ? null
               : AdaptiveBottomNavigationBar(
-                  selectedItemColor: Colors.orange,
-                  unselectedItemColor: CupertinoColors.activeBlue,
-
                   selectedIndex: widget.navigationShell.currentIndex,
                   onTap: (index) => onDestinationSelected(index, context),
                   items: [